fix(backend): enumerate achievement types in some response schema
