+++ packages/backend/src/models/json-schema/notification.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import { ACHIEVEMENT_TYPES } from '@/core/AchievementService.js';
+import { ACHIEVEMENT_TYPES } from '@/models/UserProfile.js';
 import { notificationTypes, userExportableEntities } from '@/types.js';
 
 const baseSchema = {
