+++ packages/backend/src/server/api/endpoints/users/achievements.ts
@@ -7,6 +7,7 @@ import { Inject, Injectable } from '@nestjs/common';
 import { Endpoint } from '@/server/api/endpoint-base.js';
 import type { UserProfilesRepository } from '@/models/_.js';
 import { DI } from '@/di-symbols.js';
+import { ACHIEVEMENT_TYPES } from '@/models/UserProfile.js';
 
 export const meta = {
 	requireCredential: false,
@@ -18,6 +19,7 @@ export const meta = {
 			properties: {
 				name: {
 					type: 'string',
+					enum: ACHIEVEMENT_TYPES,
 				},
 				unlockedAt: {
 					type: 'number',
